% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Box_Deconv.R
\name{Box_Deconv}
\alias{Box_Deconv}
\title{Box_Deconv}
\usage{
Function that generates a boxplot comparing two conditions for each cell type from a df and a condition vector.
}
\arguments{
\item{data}{Parameter in which the matrix resulting from the deconvolution is entered, with the cell types in the rows and the samples analysed in the columns.}

\item{cond}{Vector that assigns a condition to each sample of the data frame.}

\item{results_dir}{Parameter to enter the directory where the graphics will be generated (e.g. "/bicoh/nidia/Deconv").}

\item{f_name}{Name you want to give to the file, by default it is "Deconvolution Boxplot".}
}
\value{
The function returns the boxplots graph in the specified directory.
}
\description{
Box_Deconv
}
\examples{
Box_Deconv(data= C_abs, cond= fractions, results_dir = "/bicoh/nidia/Deconv/", f_name= "Box_CIBERSORT_abs")
}
\author{
Nidia Barco Armengol
}
