% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBERSORT_Deconv.R
\name{CIBERSORT_Deconv}
\alias{CIBERSORT_Deconv}
\title{CIBERSORT_Deconv}
\usage{
Function that generates the relative or absolute deconvolution plots with the CIBERSORT method, with the non-fractional plot, returns a df resulting from the deconvolution and two graphs of deconvolution data.
}
\arguments{
\item{matrix}{Parameter in which a matrix (in .txt, .csv or .tsv format) is introduced, which in the first column has the names of the genes and in the rest of the columns the expressions for each gene of the different samples.}

\item{sig.matrix}{Parameter in which a matrix (in .txt, .csv or .tsv format) is introduced, with the first column containing the names of the genes and the rest of the columns containing the expression signatures of different cell types.}

\item{method}{Parameter where you enter "abs" if you want to do the absolute deconvolution of the matrix, by default it is "rel".}

\item{results_dir}{Parameter to enter the directory where the graphics will be generated (e.g. "/bicoh/nidia/Deconv").}

\item{height_deconv}{Parameter in which a value is entered that determines the height of the bar chart, by default it is 10.}

\item{width_deconv}{Parameter in which a value is entered that determines the width of the bar chart, by default it is 9.}

\item{height_heatmap}{Parameter in which a value is entered that determines the height of the heatmap, by default it is 700.}

\item{width_heatmap}{Parameter in which a value is entered that determines the width of the heatmap, by default it is 700.}

\item{name}{Parameter to enter the name you want to be included in the generated graphics (Ex. "LM22" -> "CIBERSORT_Deconv_LM22_rel_plot.png"), default would be ("CIBERSORT_Deconv_(sig.matrix)_rel_plot.png").}

\item{number_format}{Parameter that allows to change the visualization of the numbers in the heatmap. ("\%.2f") for two decimals and ("\%.1e") for exponential notation.}

\item{cond}{Vector that assigns a condition to each sample of the data frame.}

\item{byCond}{Parameter in which TRUE is introduced if we introduce a vector so that the function generates the graphs dividing the samples by its condition, by default it is FALSE.}
}
\value{
Returns the generated graphs and the df of the deconvolution.
}
\description{
CIBERSORT_Deconv
}
\examples{
c<- CIBERSORT_Deconv(matrix = matrix, sig.matrix = sig.matrix, results_dir = results_dir, cibersortpath = path_ciber, by_Cond = TRUE, cond = fractions, method = "abs")
}
\author{
Nidia Barco Armengol
}
