% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution_graph.R
\name{Deconvolution_graph}
\alias{Deconvolution_graph}
\title{Deconvolution_graph}
\usage{
Function that generate the staked bar chart of different deconvolution methods.
}
\arguments{
\item{df}{Parameter in which the matrix resulting from the deconvolution is entered, with the cell types in the rows and the samples analysed in the columns.}

\item{file}{Parameter to enter the name under which you want to save the resulting graphic ("EPIC_Deconv_Exp_data_deconv.png").}

\item{results_dir}{Parameter to enter the directory where the graphics will be generated (e.g. "/bicoh/nidia/Deconv").}

\item{height}{Parameter in which a value is entered that determines the height of graph, by default it is 10.}

\item{width}{Parameter in which a value is entered that determines the width of the graph, by default it is 9.}
}
\value{
Returns the generated graph of the deconvolution.
}
\description{
Deconvolution_graph
}
\examples{
Deconvolution_graph(df= C, file = file, results_dir = results_dir, height = height_deconv, width = width_deconv)
}
\author{
Nidia Barco Armengol
}
