% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPIC_Deconv.R
\name{EPIC_Deconv}
\alias{EPIC_Deconv}
\title{EPIC_Deconv}
\usage{
Function that generates the relative deconvolution plots with the EPIC method, with the non-fractional plot, and returns a df resulting from the deconvolution.
}
\arguments{
\item{matrix}{Parameter in which a matrix (dataframe) is introduced, which in the first column has the names of the genes and in the rest of the columns the expressions for each gene of the different samples.}

\item{sig.matrix}{Parameter in which a matrix (dataframe) is introduced, with the first column containing the names of the genes and the rest of the columns containing the expression signatures of different cell types.}

\item{results_dir}{Parameter to enter the directory where the graphics will be generated (e.g. "/bicoh/nidia/Deconv").}

\item{height_deconv}{Parameter in which a value is entered that determines the height of the bar chart, by default it is 10.}

\item{width_deconv}{Parameter in which a value is entered that determines the width of the bar chart, by default it is 9.}

\item{height_heatmap}{Parameter in which a value is entered that determines the height of the heatmap, by default it is 700.}

\item{width_heatmap}{Parameter in which a value is entered that determines the width of the heatmap, by default it is 700.}

\item{name}{Parameter to enter the name you want to be included in the generated graphics (Ex. "LM22" -> "EPIC_Deconv_LM22_plot.png"), default would be ("EPIC_Deconv_plot.png").}

\item{number_format}{Parameter that allows to change the visualization of the numbers in the heatmap. ("\%.2f") for two decimals and ("\%.1e") for exponential notation.}

\item{byCond}{Parameter in which TRUE is introduced if we introduce a vector so that the function generates the graphs dividing the samples by its condition, by default it is FALSE.}

\item{cond}{Vector that assigns a condition to each sample of the data frame.}

\item{x.size}{This parameter allows you to select the font size of the X-axis in the bar chart. The default is 10.}

\item{y.size}{This parameter allows you to select the font size of the Y-axis in the bar chart. The default is 10.}

\item{l.size}{This parameter allows you to select the font size of the legend in the bar chart. The default is 10.}
}
\value{
Returns the generated graphs and the df of the deconvolution.
}
\description{
EPIC_Deconv
}
\examples{
c<- EPIC_Deconv(matrix = matrix, sig.matrix = sig.matrix, results_dir = results_dir, byCond = TRUE, cond = fractions)
}
\author{
Nidia Barco Armengol
}
