% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEDIT_functions.R
\name{GEDIT_functions}
\alias{GEDIT_functions}
\alias{A_GEDITDecon}
\title{GEDIT_functions}
\usage{
Document with the necessary functions for deconvolution with the GEDIT method.
}
\description{
GEDIT_functions
}
\author{
Nidia Barco Armengol
}
