% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heatmap_graph.R
\name{Heatmap_graph}
\alias{Heatmap_graph}
\title{Heatmap_graph}
\usage{
Function that generate a heatmap with deconvolution resultant data.
}
\arguments{
\item{df}{Parameter in which the matrix resulting from the deconvolution is entered, with the cell types in the rows and the samples analysed in the columns.}

\item{file}{Parameter to enter the name under which you want to save the resulting graphic ("EPIC_Heatmap_Exp_data_heatmap.png").}

\item{results_dir}{Parameter to enter the directory where the graphics will be generated (e.g. "/bicoh/nidia/Deconv")}

\item{number_format}{Parameter that allows to change the visualization of the numbers in the heatmap. "%.2f" for two decimals and "%.1e" for exponential notation.}

\item{width}{Parameter in which a value is entered that determines the width of the graph, by default it is 700.}

\item{height}{Parameter in which a value is entered that determines the height of the graph, by default it is 700.}
}
\value{
Returns the generated heatmap graph.
}
\description{
Heatmap_graph
}
\examples{
Heatmap_graph(df= C, file = file, results_dir = results_dir, number_format = number_format, width = width_heatmap, height = height_heatmap)
}
\author{
Nidia Barco Armengol
}
