% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mouse_Human_converter.R
\name{Mouse_Human_converter}
\alias{Mouse_Human_converter}
\title{Mouse_Human_converter}
\usage{
This function convert mouse expression data into human expression data, changing the mouse genes for her analog in human.
}
\arguments{
\item{matrix}{Parameter in which a matrix (dataframe) is introduced, which in the first column has the names of the genes and in the rest of the columns the expressions for each gene of the different samples.}

\item{resultant_file}{Parameter to enter the name we want the transformed matrix to have (Ex. "counts_mf_human.txt").}

\item{results_dir}{Parameter to enter the directory where the resultant_matrix.txt will be generated (e.g. "/bicoh/nidia/Deconv").}
}
\value{
This function returns the expression data in human into new file.
}
\description{
Mouse_Human_converter
}
\examples{
M<- Mouse_Human_converter(matrix = matrix,resultant_file = "counts_mf_human.txt", results_dir = results_dir)
}
\author{
Nidia Barco Armengol
}
