% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Preprocesment_EPIC_FERDEEP.R
\name{Preprocesment_EPIC_FARDEEP}
\alias{Preprocesment_EPIC_FARDEEP}
\title{Preprocesment_EPIC_FARDEEP}
\usage{
Function for realize the pre-procesment for EPIC and FARDEEP methods
}
\arguments{
\item{matrix}{Parameter in which a matrix (dataframe) is introduced, which in the first column has the names of the genes and in the rest of the columns the expressions for each gene of the different samples.}

\item{sig.matrix}{Parameter in which a matrix (dataframe) is introduced, with the first column containing the names of the genes and the rest of the columns containing the expression signatures of different cell types.}
}
\value{
List with processed matrices.
}
\description{
Preprocesment_EPIC_FARDEEP
}
\examples{
M<- Preprocesment_EPIC_FARDEEP(matrix = matrix, sig.matrix = sig.matrix)
}
\author{
Nidia Barco Armengol
}
